Refactor imports to use RidesCatalog context instead of Rides context across multiple components and utility functions
