+++ controller/apps/explorer/frontend/src/components/AnalysisStatusTag/index.tsx
@@ -1,6 +1,6 @@
 /* * */
 
-import { type ExtendedRideDisplay } from '@/contexts/Rides.context';
+import { type ExtendedRideDisplay } from '@/contexts/RidesCatalog.context';
 import { IconCheck, IconX } from '@tabler/icons-react';
 import { Tag } from '@tmlmobilidade/ui';
