+++ controller/apps/explorer/frontend/src/components/RidesBoard/index.tsx
@@ -5,8 +5,8 @@
 import { FlapLine } from '@/components/FlapLine';
 import { Label } from '@/components/Label';
 import { SeenStatusTag } from '@/components/SeenStatusTag';
-import { type ExtendedRideDisplay } from '@/contexts/Rides.context';
 import { useRidesBoardContext } from '@/contexts/RidesBoard.context';
+import { type ExtendedRideDisplay } from '@/contexts/RidesCatalog.context';
 
 import styles from './styles.module.css';
