+++ controller/apps/explorer/frontend/src/components/SeenStatusTag/index.tsx
@@ -1,6 +1,6 @@
 /* * */
 
-import { type ExtendedRideDisplay } from '@/contexts/Rides.context';
+import { type ExtendedRideDisplay } from '@/contexts/RidesCatalog.context';
 
 import styles from './styles.module.css';
