+++ controller/apps/explorer/frontend/src/utils/get-delay-status.ts
@@ -1,6 +1,6 @@
 /* * */
 
-import { type ExtendedRideDisplay } from '@/contexts/Rides.context';
+import { type ExtendedRideDisplay } from '@/contexts/RidesCatalog.context';
 import { type Ride } from '@tmlmobilidade/types';
 
 /**
