+++ controller/apps/explorer/frontend/src/utils/get-seen-status.ts
@@ -1,6 +1,6 @@
 /* * */
 
-import { type ExtendedRideDisplay } from '@/contexts/Rides.context';
+import { type ExtendedRideDisplay } from '@/contexts/RidesCatalog.context';
 import { type Ride } from '@tmlmobilidade/types';
 import { getUnixTimestamp } from '@tmlmobilidade/utils';
